Moved code to appropriate class
