+++ src/main/Main.java
@@ -8,7 +8,6 @@ import main.methods.CombinatoricsPermutations;
 import main.methods.Helpers;
 import main.methods.Log;
 import main.methods.ProcessInput;
-
 import java.awt.BorderLayout;
 import java.awt.Font;
 import javax.swing.JButton;
@@ -19,7 +18,6 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Timer;
 import java.util.TimerTask;
@@ -180,7 +178,7 @@ public class Main {
 					try{
 						openPrintWriter("wordList.txt");
 						List<List<String>> tempOutput = new ArrayList<List<String>>();
-						tempOutput = buildWords( ProcessInput.extractInput(txtWords.getText() , Settings.wordSeparator), Settings.wordCombinator , Settings.optionSeparator ) ;
+						tempOutput = ProcessInput.buildWords( ProcessInput.extractInput(txtWords.getText() , Settings.wordSeparator), Settings.wordCombinator , Settings.optionSeparator ) ;
 						CombinatoricsPermutations.combinations2D( Helpers.convertListToStringArray(tempOutput) );
 						closePrintWriter();
 						Log.appendToTextPanel("Test: "+test, 'I');
@@ -278,12 +276,6 @@ public class Main {
 	}
 	
 	
-	
-	
-////////////////////////////////////////////////////////////////////////////////////////
-//Working Methods\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
-////////////////////////////////////////////////////////////////////////////////////////
-
 	private static void printMemoryUsageInInterval( int intervalSeconds){
 		timer = new Timer();
 		timer.scheduleAtFixedRate(new TimerTask() {
@@ -348,245 +340,6 @@ public class Main {
 		}else
 			Settings.lowMemory = false;
 	}
-
-
-	/*
-	 * Getters	
-	 */
-	public static String getTxtWords(){
-		return txtWords.getText();
-	}
-	
-	
-		
-		
-		
-		
-		/*
-		 * Build words from options
-		 */
-		/*
-		 * NOTE:
-		 * Total number of generated words is: X * Y
-		 * Where X is number of words generated by combinations2D and Y same but from heapPermutation
-		 * Example: Input is {{"sad","SAD"},{"najebo","NAJEBO"},{"sam","SAM"},{"test","TEST", "Test", "ayy"}}
-		 * X would then be 2*2*2*4 = 32 
-		 * Y would be 4! (1*2*3*4) because we have 4 blocks (4 words that make a sentence)
-		 * Total = 768
-		 */	
-			
-		//int Y = inputWords.size(); daje nullPointer exception ako je input prazan		
-		private static List<List<String>> buildWords( ArrayList<String> inputWords , char combiningSeparator, char optionSeparator){
-			int i;
-			int X;
-			int Y = inputWords.size();
-			List<List<String>> tempOutput = new ArrayList<List<String>>();
-			
-			//Add Y words and start adding words and their options
-			for( i=0; i<Y; i++){
-				tempOutput.add(new ArrayList<String>());
-				extractWords( tempOutput, i, inputWords.get(i).toString(), combiningSeparator, optionSeparator );
-			}
-			
-			Log.writeDebug( "Base words that were extracted from input(without their options):" );
-			for( i=0; i<Y; i++){
-				for (String value : tempOutput.get(i)) {
-					Log.writeDebug( (i+1) + ":" + value );
-				}
-			}
-			
-			return tempOutput;
-		}
-		
-		
-		
-		
-		/*
-		 * Extract words from string
-		 */
-		
-		public static void extractWords( List<List<String>> inputList , int index , String extractString , char combiningSeparator, char optionSeparator ){
-			String baseString;
-			String optionString;
-			int indexCombining = extractString.indexOf(combiningSeparator);
-
-			if( indexCombining!=-1 ){
-				/*
-				 * There is combining separator in string so we will break the word
-				 * in 2 parts and call this function again with each part
-				 */
-			   Log.writeDebug("Found combining separator in string: " + extractString );
-			   String tempInput = extractString ;
-			   StringBuilder temp = new StringBuilder();
-			   for( int i=0; i<tempInput.length(); i++ ){
-				   if( tempInput.charAt(i)==combiningSeparator ){
-					   extractWords( inputList, index, temp.toString() , combiningSeparator, optionSeparator );
-					   temp = new StringBuilder();
-
-				   }else if( i==tempInput.length()-1 ){
-					   temp.append( tempInput.charAt(i) );
-					   extractWords( inputList, index, temp.toString() , combiningSeparator, optionSeparator );
-					   //temp = new StringBuilder();
-				   }else{
-					   temp.append( tempInput.charAt(i) );
-				   }
-			   }
-			   return;
-			}else{
-				/*
-				 * There is NO combining separator in word meaning word is single
-				 * Now we look if word has additional options or not
-				 */
-				int indexOption = extractString.indexOf(optionSeparator);
-				if( indexOption!=-1 ){
-					/*
-					 * There are options in the word so we will have to generate new words
-					 * based on those options and base string (which is an original word)
-					 */
-					baseString = extractString.substring(0, indexOption);
-					optionString = extractString.substring(indexOption);
-					inputList.get(index).add(baseString);
-					/*
-					 * Split option string to multiple parts then call function for every option part
-					 * We used optionSeparatorDouble because option separator is a double char in the program
-					 */
-					String optionSeparatorDouble = optionSeparator+""+optionSeparator;
-					String[] word = optionString.split(optionSeparatorDouble);
-					word = removeEmpty(word);
-					
-					Log.writeDebug("Printing option array for word: '" + extractString + "' in format (ID:option) :" );
-					for(int k =0; k<word.length;k++)
-						Log.writeDebug( k + ": " + word[k] );
-					
-						for(String optionPart: word){
-							/*
-							 * Eliminate empty words produced by optionString.split
-							 */
-							//log.append("Test:"+optionPart+"\n");
-							//if(!optionPart.isEmpty()){
-								/*
-								 * Before calling the option method first extract sub-options
-								 */
-							Log.writeDebug("Spliting:'"+optionPart+"' to suboptions!" );
-								String[] subOptions = optionPart.split( ""+optionSeparator );
-								/*
-								 * Now remove empty words from sub-options produced by optionString.split
-								 */
-								subOptions = removeEmpty(subOptions);
-								Log.writeDebug("Printing suboptions (ID:suboption) where ID=0 means base option:" );
-								for(int k =0; k<subOptions.length;k++)
-									Log.writeDebug(k+":"+subOptions[k] );
-								/*
-								 * String in first index of finalSubOptions is the option while rest are subOptions
-								 */
-								//Log.writeDebug("baseString: "+baseString+"  ,optionPart:"+optionPart+"  ,subOptions: (printed above)");
-								extractOptions( inputList, index, baseString, subOptions );
-							//}
-						}
-					return;
-				}else{
-					/*
-					 * There are no options in word so we just add it to the list
-					 */
-					Log.writeDebug("No options in word:"+extractString );
-					inputList.get(index).add(extractString);
-					return;
-				}
-			}
-			
-		}
-		
-		
-		
-		/*
-		 * Build options
-		 * optionString is already checked for empty in the previous method
-		 * String in first index of finalSubOptions is the option while rest are subOptions
-		 */
-		public static void extractOptions( List<List<String>> inputList , int index , String baseString, String[] subOptions ){
-			
-			//Potential error handling (some of this might not be needed)
-			if( !(subOptions.length>0) ) {
-				Log.write("Options have been found for word "+baseString+" but program couldn't read them!\n"
-						+ "(length!>0)" , 'W');
-				return;
-			}/*else if( subOptions.length==1 ) {
-				log.append("Options have been found for word "+baseString+" but program couldn't read them!\n"
-						+ "(length==1)\n");
-				return;
-			}*///Not needed?
-			
-			//Call functions based on found options
-			
-			switch (subOptions[0]) {
-	        case "b":
-	        	if( subOptions.length>1 ){
-	        		Log.write("Option 'b' can't have suboptions! ("+subOptions[1]+" etc...)" , 'E');
-	        	}else{
-
-	        	}
-	        	break;
-	        case "l":
-	        	if( subOptions.length>1 ){
-	        		Log.write("Option 'l' can't have suboptions! ("+subOptions[1]+" etc...)" , 'E');
-	        		break;
-	        	}else{
-	        		//inputList.get(index).add( baseString.toLowerCase() );
-	        	}
-	        	break;
-	        case "u":
-	        	if( subOptions.length>1 ){
-	        		Log.write("Option 'u' can't have suboptions! ("+subOptions[1]+" etc...)" , 'E');
-	        		break;
-	        	}else{
-	        		inputList.get(index).add( baseString.toUpperCase() );
-	        	}
-	        	break;
-	        case "ip":
-	        	Log.write("Option 'ip' is not curently in function!" , 'W');
-	        	break;
-	        case "ab":
-	        	Log.write("Option 'ab' is not curently in function!" , 'W');
-	        	break;
-	        case "ae":
-	        	Log.write("Option 'ae' is not curently in function!" , 'W');
-	        	break;
-	        default:
-	        	Log.write("Can't find option '" +subOptions[0]+ "'" , 'E');
-			}
-			
-			
-		}
-		
-		
-		/*
-		 * Eliminate empty words in string array
-		 * Words are considered empty if they are null or have 0 characters
-		 */
-		
-		public static String[] removeEmpty(String[] input){
-			List<String> list = new ArrayList<String>(Arrays.asList(input));
-			list.removeAll(Arrays.asList("", null));
-			return list.toArray(new String[0]);
-		}
-		
-		/*
-		/*
-		 * Convert List<List<String>> to String[][] 
-		 * Only temporal solution until method combinations2d is reworked so it can work with lists instead of arrays
-		 */
-		
-		public static String[][] convertListToStringArray( List<List<String>> theStrings ){
-			//List<List<String>> theStrings ... coming from somewhere
-			String[][] stringsAsArray = new String[theStrings.size()][];
-			for (int i=0; i<theStrings.size();i++) {
-			  List<String> aList = theStrings.get(i);
-			  stringsAsArray[i] = aList.toArray(new String[aList.size()]);
-			}
-			return stringsAsArray;
-		}
-		
-		
 	
 	
 }
